+++ wrapper/src/test/java/software/amazon/jdbc/ConnectionPluginChainBuilderTests.java
@@ -35,7 +35,7 @@ import org.mockito.MockitoAnnotations;
 import software.amazon.jdbc.plugin.ConnectTimeConnectionPlugin;
 import software.amazon.jdbc.plugin.DefaultConnectionPlugin;
 import software.amazon.jdbc.plugin.ExecutionTimeConnectionPlugin;
-import software.amazon.jdbc.plugin.IamAuthConnectionPlugin;
+import software.amazon.jdbc.plugin.iam.IamAuthConnectionPlugin;
 import software.amazon.jdbc.plugin.dev.DeveloperConnectionPlugin;
 import software.amazon.jdbc.plugin.efm2.HostMonitoringConnectionPlugin;
 import software.amazon.jdbc.plugin.failover.FailoverConnectionPlugin;
